some type casting to avoid warnings due to assigning double values to int.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@1863 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ framework/modules/progress/src/ProgressModule.cc
@@ -42,7 +42,7 @@ void ProgressModule::event()
 {
   ++m_evtNr;
   //Calculate the order of magnitude
-  int order = (m_evtNr == 0) ? 1 : min(log10(m_evtNr), (double)m_maxOrder);
-  int interval = pow(10., order);
+  int order = (m_evtNr == 0) ? 1 : static_cast<int>(min(log10(m_evtNr), (double)m_maxOrder));
+  int interval = static_cast<int>(pow(10., order));
   if (m_evtNr % interval == 0) B2INFO(m_output % m_runNr % m_evtNr);
 }
